Complete basic server
This commit adds the completed code for a basic server using nodejs
only and only serves GET requests.

This project also has an event emitter to track and log events. Errors
are logged and sent to the errLog.txt file in the logs directory. Request
logs are stored in the reqLog.txt file.